FIX : en passant, remove spurrious error message if there is nothing to stop.

+++ scripts/stop_dev_backends.sh
@@ -19,6 +19,8 @@
 . $(dirname $0)/backend_setup.sh
 
 for file in ${SCRIPT_DIR}/*.pid; do
-  kill $(cat $file)
-  rm $file
+  if [ -f $file ]; then
+    kill $(cat $file)
+    rm $file
+  fi
 done
